+++ library/rbxmk.go
@@ -52,7 +52,7 @@ var RBXMK = rbxmk.Library{
 }
 
 func openRBXMK(s rbxmk.State) *lua.LTable {
-	lib := s.L.CreateTable(0, 14)
+	lib := s.L.CreateTable(0, 13)
 	lib.RawSetString("cookiesFrom", s.WrapFunc(rbxmkCookiesFrom))
 	lib.RawSetString("decodeFormat", s.WrapFunc(rbxmkDecodeFormat))
 	lib.RawSetString("diffDesc", s.WrapFunc(rbxmkDiffDesc))
@@ -60,7 +60,6 @@ func openRBXMK(s rbxmk.State) *lua.LTable {
 	lib.RawSetString("formatCanDecode", s.WrapFunc(rbxmkFormatCanDecode))
 	lib.RawSetString("loadFile", s.WrapFunc(rbxmkLoadFile))
 	lib.RawSetString("loadString", s.WrapFunc(rbxmkLoadString))
-	lib.RawSetString("newAttrConfig", s.WrapFunc(rbxmkNewAttrConfig))
 	lib.RawSetString("newCookie", s.WrapFunc(rbxmkNewCookie))
 	lib.RawSetString("newDesc", s.WrapFunc(rbxmkNewDesc))
 	lib.RawSetString("patchDesc", s.WrapFunc(rbxmkPatchDesc))
@@ -180,12 +179,6 @@ func rbxmkLoadString(s rbxmk.State) int {
 	}
 	s.L.Push(fn)
 	return 1
-}
-
-func rbxmkNewAttrConfig(s rbxmk.State) int {
-	var v rtypes.AttrConfig
-	v.Property = string(s.PullOpt(1, "string", types.String("")).(types.String))
-	return s.Push(&v)
 }
 
 func rbxmkNewCookie(s rbxmk.State) int {
@@ -462,16 +455,6 @@ func dumpRBXMK(s rbxmk.State) dump.Library {
 					CanError:    true,
 					Summary:     "Libraries/rbxmk:Fields/loadString/Summary",
 					Description: "Libraries/rbxmk:Fields/loadString/Description",
-				},
-				"newAttrConfig": dump.Function{
-					Parameters: dump.Parameters{
-						{Name: "property", Type: dt.Prim("string")},
-					},
-					Returns: dump.Parameters{
-						{Type: dt.Prim("AttrConfig")},
-					},
-					Summary:     "Libraries/rbxmk:Fields/newAttrConfig/Summary",
-					Description: "Libraries/rbxmk:Fields/newAttrConfig/Description",
 				},
 				"newCookie": dump.Function{
 					Parameters: dump.Parameters{