+++ reflect/AttrConfig.go
@@ -33,6 +33,29 @@ func AttrConfig() rbxmk.Reflector {
 				},
 			},
 		},
+		Constructors: rbxmk.Constructors{
+			"new": rbxmk.Constructor{
+				Func: func(s rbxmk.State) int {
+					var v rtypes.AttrConfig
+					v.Property = string(s.PullOpt(1, "string", types.String("")).(types.String))
+					return s.Push(&v)
+				},
+				Dump: func() dump.MultiFunction {
+					return dump.MultiFunction{
+						dump.Function{
+							Parameters: dump.Parameters{
+								{Name: "property", Type: dt.Prim("string")},
+							},
+							Returns: dump.Parameters{
+								{Type: dt.Prim("AttrConfig")},
+							},
+							Summary:     "Types/AttrConfig:Constructors/new/Summary",
+							Description: "Types/AttrConfig:Constructors/new/Description",
+						},
+					}
+				},
+			},
+		},
 		Dump: func() dump.TypeDef {
 			return dump.TypeDef{
 				Summary:     "Types/AttrConfig:Summary",